requirements.txt
numpy
torch
Pillow
torchvision
pytorch-lightning
pycocotools
